Move SchemaRegistryConfigAttribute to the appropriate package
